wrapped up controller, moved all userServices to its own directory.
Deletion logic still returns a silent error, however it deletes the user. Will come back to this at another point.
